requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
flask-mongoengine==1.0.0
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
WTForms==3.0.1
zipp==3.17.0
